migrate pkg name and groupId from s.a.i to c.a.i
