+++ src/main/java/com/amazon/ionpathextraction/PathExtractor.java
@@ -11,7 +11,7 @@
  * language governing permissions and limitations under the License.
  */
 
-package software.amazon.ionpathextraction;
+package com.amazon.ionpathextraction;
 
 import com.amazon.ion.IonReader;
