+++ src/main/java/com/amazon/ionpathextraction/internal/Annotations.java
@@ -11,7 +11,7 @@
  * language governing permissions and limitations under the License.
  */
 
-package software.amazon.ionpathextraction.internal;
+package com.amazon.ionpathextraction.internal;
 
 import java.util.stream.IntStream;
