+++ src/main/java/com/amazon/ionpathextraction/internal/Preconditions.java
@@ -11,9 +11,9 @@
  * language governing permissions and limitations under the License.
  */
 
-package software.amazon.ionpathextraction.internal;
+package com.amazon.ionpathextraction.internal;
 
-import software.amazon.ionpathextraction.exceptions.PathExtractionException;
+import com.amazon.ionpathextraction.exceptions.PathExtractionException;
 
 /**
  * <p>
