+++ src/test/kotlin/com/amazon/ionpathextraction/ExampleTest.java
@@ -11,7 +11,7 @@
  * language governing permissions and limitations under the License.
  */
 
-package software.amazon.ionpathextraction;
+package com.amazon.ionpathextraction;
 
 import static org.junit.Assert.assertEquals;
